move main stack infrastructure into /infrastructure/instance-scheduler
