+++ source/infrastructure/instance-scheduler/lib/app-registry.ts
