+++ source/infrastructure/pipeline/cdk.json
@@ -1,0 +1,4 @@
+{
+  "app": "npx ts-node bin/testing-pipeline.ts",
+  "output": "../../../build/cdk.pipeline.out"
+}
