+++ source/infrastructure/pipeline/lib/code-star-source.ts
