+++ source/infrastructure/tsconfig.json
@@ -31,6 +31,6 @@
   ],
   "exclude": [
     "node_modules",
-    "jest.config.ts"
+    "**/jest.config.ts"
   ],
 }
